Securisation de la suppression

+++ src/Controller/ResourceController.php
@@ -5,25 +5,36 @@ namespace App\Controller;
 use App\Core\ResourceManager;
 use App\Entity\Resource;
 use App\Form\ResourceType;
+use App\Security\ResourceVoter;
 use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\BinaryFileResponse;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route("/resource", name: "resource_")]
-class ResourceController extends AbstractController
+class ResourceController extends CortestAbstractController
 {
 
     #[Route("/download/{id}", name: "download")]
     public function download(
-        ResourceManager $resourceManager,
-        Resource        $resource): BinaryFileResponse
+        ResourceManager        $resourceManager,
+        EntityManagerInterface $entityManager,
+        Resource               $resource): Response
     {
-        $filePath = $resourceManager->resourceFilePath($resource);
-        return $this->file(file: $filePath, fileName: $resource->file_nom);
+        $filePath = $resourceManager->resourcefilePathOrNull($resource);
+
+        if ($filePath == null) {
+            $this->addFlash("danger", "Le fichier n'existe pas. Suppression de la resource");
+
+            $entityManager->remove($resource);
+            $entityManager->flush();
+
+            return $this->redirectToRoute("home");
+        } else {
+            return $this->file(file: $filePath, fileName: $resource->file_nom);
+        }
     }
 
     #[Route("/creer", name: "creer")]
@@ -31,7 +42,7 @@ class ResourceController extends AbstractController
                           ResourceManager        $resourceManager,
                           EntityManagerInterface $entityManager): Response
     {
-        $resource = new Resource(id: 0, nom: "", file_nom: "", user: $this->getUser());
+        $resource = new Resource(id: 0, nom: "", file_nom: "", user: $this->getNonNullUser());
         $form = $this->createForm(ResourceType::class, $resource);
 
         $form->handleRequest($request);
@@ -41,9 +52,16 @@ class ResourceController extends AbstractController
             $entityManager->persist($resource);
             $entityManager->flush();
 
-            $resourceManager->upload($form->get("file")->getData(), $resource);
+            $result = $resourceManager->upload($form->get("file")->getData(), $resource);
 
-            return $this->redirectToRoute("home");
+            if (!$result) {
+                $entityManager->remove($resource);
+                $entityManager->flush();
+                $this->addFlash("danger", "Echec de la mise en ligne du fichier");
+            } else {
+                $this->addFlash("success", "Resource enregistrée");
+                return $this->redirectToRoute("home");
+            }
         }
 
         return $this->render("resource/form_creer.html.twig", ["form" => $form->createView()]);
@@ -56,12 +74,14 @@ class ResourceController extends AbstractController
         Resource               $resource
     ): RedirectResponse
     {
+        $this->denyAccessUnlessGranted(attribute: ResourceVoter::DELETE, subject: $resource);
+
         $resourceManager->delete($resource);
 
         $entityManager->remove($resource);
         $entityManager->flush();
 
-        $this->addFlash("info", "Resource supprimée");
+        $this->addFlash("success", "Resource supprimée");
 
         return $this->redirectToRoute("home");
     }
